pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "rnse"
version = "0.1.0"
description = "Report Number Synchronisation Engine — deterministic placeholder substitution for Word reports"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "openpyxl>=3.1",
    "python-docx>=1.1",
    "rich>=13.0",
]

[project.scripts]
rnse = "rnse.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
